Add support for defining which model fields are to be excluded from the index by defining those fields directly on the model in question.
Previously we introduced new "exclude_from_index" function argument
which allows developer to decide which fields to include from index
when calling the translator library.

That abstraction is not ideal though since it means that each caller
of the translator library needs to be aware which fields should be
excluded from index and that breaks separation of concern.

In majority of scenarios, this list of fields is static and should be
defined at the source - aka directly on the model in question.

This way other pieces of code doesn't need to be aware of any of the
model specifics and those are completely encapsulated on the model.